Removed the check for lift segments
This only runs after the lift event listener, so we don't need to check for segments. However, we do need to change the empty segments to an object from "undefined" so that we can check the negation option against a person with no segments yet. Bonus - a little shorter code.